Split Nek5000 into Nek5000 and Nektools
